requirements.txt
dash
plotly
paho-mqtt
pandas
dash_bootstrap_components